Separate query parameters and variables vue files
